Replace INFINITY sentinels with finite SDF_FAR, make octree LoD fields uint8_t
The stored SDF field used INFINITY as the 'no data / far-empty' sentinel
(untouched cells, pruned subtrees, out-of-domain samples). Shape functions
themselves never return infinity, so mixing finite corners with the
infinite sentinel on map-edge/coarse cells produced t = INF/INF = NaN in
SDF::getPosition, corrupting QEF vertices (measured ~8,999 NaN vertex
writes per run). Replace the sentinel with a finite SDF_FAR = 1e5f so every
stored value is interpolatable and sign-consistent, and skip crossings
against sentinel corners in getPosition/getNormalFromPosition.

LoD fields lod/chunkLod were int8_t with -1 as 'unset'. The allocator
hands out raw (malloc'd/recycled) memory without running constructors, so
stale -1/garbage values survived init() if not reset. Convert both to
uint8_t with a +1 shift (0 = unset, stored = ladder level + 1): the walker
and propagation compare stored-to-stored, and only boundary consumers
(renderer LADDER levels, maxChunkLod count) decode back by -1.

+++ space/Processor.hpp
@@ -22,7 +22,7 @@ public:
     // ladder level (the node's chunkLod), the node itself and the node's own
     // root-descended bounding cube (used for the tessellation bounds — no
     // reconstructed cubes), plus the node's geometry.
-    std::function<void(int level, const OctreeNode* node, const BoundingCube &cube, Geometry &geometry, int lod)> onGeometry;
+    std::function<void(int level, const OctreeNode* node, const BoundingCube &cube, Geometry &geometry, uint8_t lod)> onGeometry;
     bool iterate(const Octree &tree, OctreeNodeData &params) override;
     void getOrder(const Octree &tree, OctreeNodeData &params, uint8_t order[8]) override;
     void virtualize(Octree * tree, const BoundingCube &cube, float * sdf, uint level, uint levels);
